final v1.3

+++ src/main/java/com/mycompany/shop/product/Product.java
@@ -4,6 +4,7 @@ import com.mycompany.shop.image.Image;
 import com.mycompany.shop.techspec.TechSpec;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -16,7 +17,7 @@ public class Product {
 
     private String name;
 
-    private Double rating;
+    private BigDecimal rating;
 
     private Integer popularity;
 
@@ -38,7 +39,7 @@ public class Product {
     @OneToMany(cascade = CascadeType.ALL, mappedBy = "product", fetch = FetchType.EAGER)
     private Set<TechSpec> techSpec = new HashSet<>();
 
-    public Product(String name, Double rating, Integer popularity, String config, Integer price,
+    public Product(String name, BigDecimal rating, Integer popularity, String config, Integer price,
                    String img, String description, String brand, String category) {
         this.name = name;
         this.rating = rating;
@@ -51,7 +52,7 @@ public class Product {
         this.category = category;
     }
 
-    public Product(Long id, String name, Double rating, Integer popularity, String config, Integer price, String img,
+    public Product(Long id, String name, BigDecimal rating, Integer popularity, String config, Integer price, String img,
                    String description, String brand, String category) {
         this.id = id;
         this.name = name;
@@ -84,11 +85,11 @@ public class Product {
         this.name = name;
     }
 
-    public Double getRating() {
+    public BigDecimal getRating() {
         return rating;
     }
 
-    public void setRating(Double rating) {
+    public void setRating(BigDecimal rating) {
         this.rating = rating;
     }
